Improve route security

+++ App/Controllers/PatientController.php
@@ -127,7 +127,12 @@ class PatientController extends Controller
 
         $em = $this->getEntityManager();
         $patientRepository = $em->getRepository(Patient::class);
-        $patient = $patientRepository->find($this->getRouteParams()['id']);
+        $id = $this->getRouteParams()['id'];
+        $patient = $patientRepository->find($id);
+        
+        if(!isset($patient))
+            throw new \Exception("Paciente $id no encontrado.", '404');
+
         $patient->delete();
         $em->flush();
 
@@ -141,7 +146,12 @@ class PatientController extends Controller
 
         $em = $this->getEntityManager();
         $patientRepository = $em->getRepository(Patient::class);
-        $patient = $patientRepository->find($this->getRouteParams()['id']);
+        $id = $this->getRouteParams()['id'];
+        $patient = $patientRepository->find($id);
+        
+        if(!isset($patient))
+            throw new \Exception("Paciente $id no encontrado.", '404');
+
         $patient->activate();
         $em->flush();
 
@@ -154,7 +164,12 @@ class PatientController extends Controller
         $this->denyAccessUnlessOneGranted(array('paciente_show', 'datosDemograficos_show'));
 
         $em = $this->getEntityManager();
-        $patient = $em->getRepository(Patient::class)->find($this->getRouteParams()['id']);
+        $id = $this->getRouteParams()['id'];
+        $patient = $em->getRepository(Patient::class)->find($id);
+
+        if(!isset($patient))
+            throw new \Exception("Paciente $id no encontrado.", '404');
+
         $this->render('Patients/patientProfile.html.twig', ['patient' => $patient, 'patientFields' => $this->getPatientFields()]);
     }
